software/s223ToMo/requirements.txt
rdflib
jinja2
pyyaml
